Parallelize the implementation to download artificats and improve performance for downloading large models.
