+++ mlflow/store/artifact/artifact_repo.py
@@ -2,11 +2,19 @@ import os
 import posixpath
 import tempfile
 from abc import abstractmethod, ABCMeta
+from collections import namedtuple
+from concurrent.futures import ThreadPoolExecutor
 
 from mlflow.utils.validation import path_not_unique, bad_path_message
 
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE, RESOURCE_DOES_NOT_EXIST
+
+# Constants used to determine max level of parallelism to use while uploading/downloading artifacts.
+# Max threads to use for parallelism.
+_NUM_MAX_THREADS = 8
+# Default number of CPUs to assume on the machine if unavailable to fetch it using os.cpu_count()
+_NUM_DEFAULT_CPUS = _NUM_MAX_THREADS / 2
 
 
 class ArtifactRepository:
@@ -19,6 +27,12 @@ class ArtifactRepository:
 
     def __init__(self, artifact_uri):
         self.artifact_uri = artifact_uri
+        # Limit the number of threads used for artifact uploads/downloads. Use at most
+        # constants._NUM_MAX_THREADS threads or 2 * the number of CPU cores available on the
+        # system (whichever is smaller)
+        num_cpus = os.cpu_count() or _NUM_DEFAULT_CPUS
+        num_artifact_workers = min(num_cpus * 2, _NUM_MAX_THREADS)
+        self.thread_pool = ThreadPoolExecutor(max_workers=num_artifact_workers)
 
     @abstractmethod
     def log_artifact(self, local_file, artifact_path=None):
@@ -104,14 +118,23 @@ class ArtifactRepository:
 
         :return: Absolute path of the local filesystem location containing the desired artifacts.
         """
+        # Represents an in-progress file artifact download to a local filesystem location
+        InflightDownload = namedtuple(
+            "InflightDownload",
+            [
+                # The artifact path, given relative to the repository's artifact root location
+                "src_artifact_path",
+                # The local filesystem destination path to which artifacts are being downloaded
+                "dst_local_path",
+                # A future representing the artifact download operation
+                "download_future",
+            ],
+        )
 
-        # TODO: Probably need to add a more efficient method to stream just a single artifact
-        #       without downloading it, or to get a pre-signed URL for cloud storage.
-        def download_artifact(src_artifact_path, dst_local_dir_path):
+        def async_download_artifact(src_artifact_path, dst_local_dir_path):
             """
             Download the file artifact specified by `src_artifact_path` to the local filesystem
             directory specified by `dst_local_dir_path`.
-
             :param src_artifact_path: A relative, POSIX-style path referring to a file artifact
                                       stored within the repository's artifact root location.
                                       `src_artifact_path` should be specified relative to the
@@ -123,16 +146,50 @@ class ArtifactRepository:
                                        subdirectories.
             :return: A local filesystem path referring to the downloaded file.
             """
+            inflight_downloads = []
             local_destination_file_path = self._create_download_destination(
                 src_artifact_path=src_artifact_path, dst_local_dir_path=dst_local_dir_path
             )
-            self._download_file(
-                remote_file_path=src_artifact_path, local_path=local_destination_file_path
+            download_future = self.thread_pool.submit(
+                self._download_file,
+                remote_file_path=src_artifact_path,
+                local_path=local_destination_file_path,
             )
-            return local_destination_file_path
+            inflight_downloads.append(
+                InflightDownload(
+                    src_artifact_path=src_artifact_path,
+                    dst_local_path=local_destination_file_path,
+                    download_future=download_future,
+                )
+            )
+            return inflight_downloads
 
-        def download_artifact_dir(src_artifact_dir_path, dst_local_dir_path):
+        def async_download_artifact_dir(src_artifact_dir_path, dst_local_dir_path):
+            """
+            Initiate an asynchronous download of the artifact directory specified by
+            `src_artifact_dir_path` to the local filesystem directory specified by
+            `dst_local_dir_path`.
+
+            This implementation is adapted from
+            https://github.com/mlflow/mlflow/blob/a776b54fa8e1beeca6a984864c6375e9ed38f8c0/mlflow/
+            store/artifact/artifact_repo.py#L93.
+
+            :param src_artifact_dir_path: A relative, POSIX-style path referring to a directory of
+                                          of artifacts stored within the repository's artifact root
+                                          location. `src_artifact_dir_path` should be specified
+                                          relative to the repository's artifact root location.
+            :param dst_local_dir_path: Absolute path of the local filesystem destination directory
+                                       to which to download the specified artifact directory. The
+                                       downloaded artifacts may be written to a subdirectory of
+                                       `dst_local_dir_path` if `src_artifact_dir_path` contains
+                                       subdirectories.
+            :return: A tuple whose first element is the destination directory of the downloaded
+                     artifacts on the local filesystem and whose second element is a list of
+                     `InflightDownload` objects, each of which represents an inflight asynchronous
+                     download operation for a file in the specified artifact directory.
+            """
             local_dir = os.path.join(dst_local_dir_path, src_artifact_dir_path)
+            inflight_downloads = []
             dir_content = [  # prevent infinite loop, sometimes the dir is recursively included
                 file_info
                 for file_info in self.list_artifacts(src_artifact_dir_path)
@@ -144,15 +201,16 @@ class ArtifactRepository:
             else:
                 for file_info in dir_content:
                     if file_info.is_dir:
-                        download_artifact_dir(
+                        inflight_downloads += async_download_artifact_dir(
                             src_artifact_dir_path=file_info.path,
                             dst_local_dir_path=dst_local_dir_path,
-                        )
+                        )[1]
                     else:
-                        download_artifact(
-                            src_artifact_path=file_info.path, dst_local_dir_path=dst_local_dir_path
+                        inflight_downloads += async_download_artifact(
+                            src_artifact_path=file_info.path,
+                            dst_local_dir_path=dst_local_dir_path,
                         )
-            return local_dir
+            return local_dir, inflight_downloads
 
         if dst_path is None:
             dst_path = tempfile.mkdtemp()
@@ -177,11 +235,39 @@ class ArtifactRepository:
 
         # Check if the artifacts points to a directory
         if self._is_directory(artifact_path):
-            return download_artifact_dir(
+            dst_local_path, inflight_downloads = async_download_artifact_dir(
                 src_artifact_dir_path=artifact_path, dst_local_dir_path=dst_path
             )
         else:
-            return download_artifact(src_artifact_path=artifact_path, dst_local_dir_path=dst_path)
+            inflight_downloads = async_download_artifact(
+                src_artifact_path=artifact_path, dst_local_dir_path=dst_path
+            )
+            assert (
+                len(inflight_downloads) == 1
+            ), "Expected one inflight download for a file artifact, got {} downloads".format(
+                len(inflight_downloads)
+            )
+            dst_local_path = inflight_downloads[0].dst_local_path
+
+        # Join futures to ensure that all artifacts have been downloaded prior to returning
+        failed_downloads = {}
+        for inflight_download in inflight_downloads:
+            try:
+                inflight_download.download_future.result()
+            except Exception as e:
+                failed_downloads[inflight_download.src_artifact_path] = repr(e)
+
+        if len(failed_downloads) > 0:
+            raise MlflowException(
+                message=(
+                    "The following failures occurred while downloading one or more"
+                    " artifacts from {artifact_root}: {failures}".format(
+                        artifact_root=self.artifact_uri,
+                        failures=failed_downloads,
+                    )
+                )
+            )
+        return dst_local_path
 
     @abstractmethod
     def _download_file(self, remote_file_path, local_path):
